feat: rm zk dependency in standalone mode (#486)
* feat: create db and table in standalone mode
