+++ src/nameserver/name_server_impl.h
@@ -34,6 +34,7 @@
 #include "client/ns_client.h"
 #include "client/tablet_client.h"
 #include "codec/schema_codec.h"
+#include "nameserver/cluster_info.h"
 #include "proto/name_server.pb.h"
 #include "proto/tablet.pb.h"
 #include "zk/dist_lock.h"
@@ -82,45 +83,6 @@ struct NearLineTabletInfo : public EndpointInfo {
     std::shared_ptr<::openmldb::client::NearLineTabletClient> client_;
 };
 
-class ClusterInfo {
- public:
-    explicit ClusterInfo(const ::openmldb::nameserver::ClusterAddress& cdp);
-
-    void CheckZkClient();
-
-    void UpdateNSClient(const std::vector<std::string>& children);
-
-    int Init(std::string& msg);  // NOLINT
-
-    bool CreateTableRemote(const ::openmldb::api::TaskInfo& task_info,
-                           const ::openmldb::nameserver::TableInfo& table_info,
-                           const ::openmldb::nameserver::ZoneInfo& zone_info);
-
-    bool DropTableRemote(const ::openmldb::api::TaskInfo& task_info, const std::string& name, const std::string& db,
-                         const ::openmldb::nameserver::ZoneInfo& zone_info);
-
-    bool AddReplicaClusterByNs(const std::string& alias, const std::string& zone_name, const uint64_t term,
-                               std::string& msg);  // NOLINT
-
-    bool RemoveReplicaClusterByNs(const std::string& alias, const std::string& zone_name, const uint64_t term,
-                                  int& code,          // NOLINT
-                                  std::string& msg);  // NOLINT
-
-    bool UpdateRemoteRealEpMap();
-
-    std::shared_ptr<::openmldb::client::NsClient> client_;
-    std::map<std::string, std::map<std::string, std::vector<TablePartition>>> last_status;
-    ::openmldb::nameserver::ClusterAddress cluster_add_;
-    uint64_t ctime_;
-    std::atomic<ClusterStatus> state_;
-    std::shared_ptr<std::map<std::string, std::string>> remote_real_ep_map_;
-
- private:
-    std::shared_ptr<ZkClient> zk_client_;
-    uint64_t session_term_;
-    // todo :: add statsus variable show replicas status
-};
-
 // the container of tablet
 typedef std::map<std::string, std::shared_ptr<TabletInfo>> Tablets;
 typedef std::map<std::string, std::shared_ptr<::openmldb::nameserver::TableInfo>> TableInfos;
@@ -162,6 +124,10 @@ class NameServerImpl : public NameServer {
     NameServerImpl(const NameServerImpl&) = delete;
 
     NameServerImpl& operator=(const NameServerImpl&) = delete;
+
+    inline bool IsClusterMode() const {
+        return startup_mode_ == ::openmldb::type::StartupMode::kCluster;
+    }
 
     void DeleteOPTask(RpcController* controller, const ::openmldb::api::DeleteTaskRequest* request,
                       ::openmldb::api::GeneralResponse* response, Closure* done);
@@ -763,6 +729,10 @@ class NameServerImpl : public NameServer {
 
     void DropProcedureOnTablet(const std::string& db_name, const std::string& sp_name);
 
+    bool AllocateTableId(uint32_t* id);
+
+    uint64_t GetTerm() const;
+
  private:
     std::mutex mu_;
     Tablets tablets_;
@@ -811,6 +781,7 @@ class NameServerImpl : public NameServer {
     std::map<std::string, std::string> sdk_endpoint_map_;
     std::unordered_map<std::string, std::unordered_map<std::string, std::vector<std::string>>> db_sp_table_map_;
     std::unordered_map<std::string, std::unordered_map<std::string, std::vector<std::string>>> db_table_sp_map_;
+    ::openmldb::type::StartupMode startup_mode_;
 };
 
 }  // namespace nameserver
